Refactor of schedule related preparer
