+++ src/main/java/org/ilgcc/app/utils/SubmissionUtilities.java
@@ -147,61 +147,4 @@ public class SubmissionUtilities {
             .toList();
   }
 
-  public static Optional<List<DayOfWeekOption>> getDaysOfWeekField(Map<String, Object> inputData, String inputName) {
-    Object value = inputData.get(inputName);
-    if (value == null) {
-      return Optional.empty();
-    } else if (value instanceof List<?> valueList) {
-      return Optional.of(valueList.stream().map(d -> DayOfWeekOption.valueOf(d.toString())).toList());
-    } else {
-      throw new IllegalArgumentException("Weekdays field does not contain a list");
-    }
-  }
-
-  public static Optional<HourlySchedule> getHourlySchedule(
-          Submission submission, String inputName, String weeklyScheduleInputName) {
-    return getHourlySchedule(submission.getInputData(), inputName, weeklyScheduleInputName);
-  }
-
-  public static Optional<HourlySchedule> getHourlySchedule(
-          Map<String, Object> inputData, String inputName, String weeklyScheduleInputName) {
-    Optional<List<DayOfWeekOption>> daysActive = getDaysOfWeekField(inputData, weeklyScheduleInputName);
-    if (daysActive.isEmpty()) {
-      return Optional.empty();
-    }
-
-    List<String> sameEveryDayField = getOptionalListField(
-            inputData, "%sHoursSameEveryDay[]".formatted(inputName), Object::toString).orElse(List.of());
-    boolean sameEveryDay = !sameEveryDayField.isEmpty() && sameEveryDayField.get(0).equalsIgnoreCase("Yes");
-
-    if (sameEveryDay) {
-      LocalTimeRange scheduleEveryDay = getTimeRangeInput(inputData, inputName, "AllDays").orElseThrow();
-      return Optional.of(new ConsistentHourlySchedule(daysActive.get(), scheduleEveryDay));
-    } else {
-      var dailyScheduleMap = daysActive.get().stream().collect(
-              toImmutableMap(
-                      identity(),
-                      day -> getTimeRangeInput(inputData, inputName, day.name()).orElseThrow()));
-      return Optional.of(new PerDayHourlySchedule(dailyScheduleMap));
-    }
-  }
-
-  public static <T> Optional<List<T>> getOptionalListField(
-          Map<String, Object> inputData, String fieldName, Function<Object, T> converter) {
-    Object value = inputData.get(fieldName);
-    if (value== null) {
-      return Optional.empty();
-    } else if (value instanceof List<?> valueList) {
-      return Optional.of(valueList.stream().map(converter).toList());
-    } else {
-      throw new IllegalArgumentException("List field does not contain a list");
-    }
-  }
-
-  public static void putSingleFieldResult(Map<String, SubmissionField> results, String fieldName, String value) {
-    results.put(fieldName, new SingleField(fieldName, value, null));
-  }
-  public static void putSingleFieldResult(Map<String, SubmissionField> results, String fieldName, String value, Integer iteration) {
-    results.put(fieldName+"_"+iteration, new SingleField(fieldName, value, iteration));
-  }
 }
